renderPlaygroundPage - cdnUrl config option (#845)
* renderPlaygroundPage - cdnUrl config option

Allow the cdn url to be passed in to renderPlaygroundPage as an option. This is to support graphql-playground being used offline by tools such as Apollo Server. See https://github.com/apollographql/apollo-server/issues/1421

* don't insert @ in cdn url if no version given

If version is not passed then don't append @ to the cdn url.

This is to aid serving from localhost from a folder with no given version number. See #845