fix: X-Apollo-Tracing No Schema Issue (#1112)
Schema fetch requests fail whenever `X-Apollo-Tracing` is included in requests to servers configured to allow certain request headers. This happens because the `X-Apollo-Tracing` header is hard-coded into all schemaFetch requests, even if the header isn't included in the allowed headers in the response to the initial pre-flight request. This causes the docs and schema to not be displayed. 

To resolve, this commit makes the following changes:
- Add  `tracing.tracingSupported` setting that when false removes `x-apollo-tracing` header from schemaFetch requests
- Don't send `x-apollo-tracing` header if `initialSchemaFetch ` is `true`